Split renderer into shape renderers
